fix the issue where the lookup join result is incorrect.

+++ core/src/operator/join/lookup_join.h
@@ -78,7 +78,7 @@ private:
     bool currentProbePositionProducedRow;
     JoinHashTables *hashTables;
     JoinProbe *joinProbe;
-    int32_t partitionedJoinPosition; // the addressIndex combined partition for build, it is encoded by ((addressIndex
+    int64_t partitionedJoinPosition; // the addressIndex combined partition for build, it is encoded by ((addressIndex
     // << shiftSize) | partition)
     std::unique_ptr<LookupJoinOutputBuilder> outputBuilder;
 };
